app/pyproject.toml
[tool.poetry]
name = "transcribe"
version = "0.1.0"
description = ""
authors = ["oCa <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.8.2"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
requests = "^2.32.3"
speechrecognition = "^3.10.4"
pydub = "^0.25.1"
googletrans = "4.0.0rc1"
python-magic = "^0.4.27"
langdetect = "1.0.9"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.139.0"
jinja2 = "^3.1.4"
aiofiles = "^23.2.1"
aiogoogle = "^5.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"